extract markdown-it from components to utility and add some related plugins
