+++ apps/shared/shared/types/module.d.ts
@@ -1,3 +1,4 @@
 declare module '#layers/shared/app/assets/external/markdown-it-span'
 declare module 'markdown-it-task-lists'
 declare module 'markdown-it-block-embed'
+declare module 'markdown-it-ins'
