+++ apps/web/package.json
@@ -39,8 +39,11 @@
     "fflate": "0.8.2",
     "lucide-vue-next": "0.548.0",
     "markdown-it": "14.1.0",
+    "markdown-it-attrs": "4.2.0",
     "markdown-it-block-embed": "0.0.3",
+    "markdown-it-ins": "4.0.0",
     "markdown-it-link-attributes": "4.0.1",
+    "markdown-it-multimd-table": "4.2.3",
     "markdown-it-task-lists": "2.1.1",
     "mupdf": "1.26.4",
     "nuxt": "4.2.0",
@@ -68,6 +71,7 @@
     "@iconify-json/prime": "1.2.4",
     "@nuxt/eslint": "1.9.0",
     "@types/markdown-it": "14.1.2",
+    "@types/markdown-it-attrs": "4.1.3",
     "@types/markdown-it-link-attributes": "3.0.5",
     "@types/node": "24.9.1",
     "@types/regexp.escape": "2.0.0",
